+++ semantic_search/test_procedures/default.json
@@ -12,8 +12,9 @@
                 "plugins": {
                   "ml_commons": {
                     "only_run_on_ml_node": "false",
-                    "model_access_control_enabled": "false",
-                    "native_memory_threshold": "99"
+                    "native_memory_threshold": "99",
+                    "allow_registering_model_via_local_file": "true",
+                    "allow_registering_model_via_url": "true"
                   }
                 }
               }
@@ -21,37 +22,28 @@
           }
         },
         {
-          "operation": {
-            "operation-type": "delete-index"
-          }
+          "operation": "delete-index"
         },
         {
-          "operation": {
-            "operation-type": "delete-ingest-pipeline",
-            "include-in-results_publishing": false
-          }
+          "operation": "delete-ingest-pipeline"
         },
         {
           "operation": {
             "operation-type": "delete-ml-model",
-            "model_name": "{{ model_name | default('huggingface/sentence-transformers/all-mpnet-base-v2')}}",
-            "include-in-results_publishing": false
+            "model-name": "{{ model_name | default('huggingface/sentence-transformers/all-mpnet-base-v2')}}"
           }
         },
         {
           "operation": {
             "operation-type": "register-ml-model",
-            "model_name": "{{ model_name | default('huggingface/sentence-transformers/all-mpnet-base-v2')}}",
-            "model_version": "{{ model_version | default('1.0.1') }}",
-            "model_format": "{{ model_format | default('TORCH_SCRIPT') }}",
-            "include-in-results_publishing": false
+            "model-name": "{{ model_name | default('huggingface/sentence-transformers/all-mpnet-base-v2')}}",
+            "model-version": "{{ model_version | default('1.0.1') }}",
+            "model-format": "{{ model_format | default('TORCH_SCRIPT') }}",
+            "model-config-file": "{{ model_config_file | default('') }}"
           }
         },
         {
-          "operation": {
-            "operation-type": "deploy-ml-model",
-            "include-in-results_publishing": false
-          }
+          "operation": "deploy-ml-model"
         },
         {
           "operation": "create-ingest-pipeline"
@@ -69,7 +61,7 @@
           "name": "check-cluster-health",
           "operation": {
             "operation-type": "cluster-health",
-            "index": "msmarco",
+            "index": "treccovid",
             "request-params": {
               "wait_for_status": "{{cluster_health | default('green')}}",
               "wait_for_no_relocating_shards": "true"
