+++ semantic_search/workload.json
@@ -2,23 +2,23 @@
 
 {
   "version": 2,
-  "description": "MS MARCO is a collection of datasets focused on deep learning in search",
+    "description": "Trec-Covid is a dataset collection of documents about COVID-19 information.",
   "indices": [
     {
-      "name": "msmarco",
+      "name": "treccovid",
       "body": "index.json"
     }
   ],
   "corpora": [
     {
-      "name": "msmarco",
-      "base-url": "https://vesa-oswl.s3.us-west-2.amazonaws.com",
+      "name": "treccovid",
+      "base-url": "https://vesa-oswl.s3.us-west-2.amazonaws.com/treccovid",
       "documents": [
         {
           "source-file": "documents.json.bz2",
-          "document-count": 8841325,
-          "compressed-bytes": 992998188,
-          "uncompressed-bytes": 3203032612
+          "document-count": 129192,
+          "compressed-bytes": 51187469,
+          "uncompressed-bytes": 211980208
         }
       ]
     }
