+++ semantic_search/workload_queries.json
@@ -1,0 +1,6 @@
+{
+  "base-url": "https://vesa-oswl.s3.us-west-2.amazonaws.com/treccovid",
+  "source-file": "queries.json.bz2",
+  "compressed-bytes": 4310,
+  "uncompressed-bytes": 16552
+}
