--- vectorsearch_embedding/workload_queries.json
@@ -1,6 +1,0 @@
-{
-  "base-url": "https://vesa-oswl.s3.us-west-2.amazonaws.com",
-  "source-file": "queries.json.bz2",
-  "compressed-bytes": 322353,
-  "uncompressed-bytes": 1078227
-}
